friend routes and components for adding and removing

+++ capstoneBE/server/db/friends.js
@@ -34,17 +34,35 @@ const fetchAllFriends = async () => {
     console.log(error);
   }
 };
+// fetch all friends that are not already friends with the user
+const fetchAvailableFriends = async (userId) => {
+  try {
+    const { rows } = await client.query(
+      `
+      SELECT u.username, u.id
+      FROM users u
+      LEFT JOIN friends f1 ON u.id = f1.friend_id AND f1.user_id = $1
+      LEFT JOIN friends f2 ON u.id = f2.user_id AND f2.friend_id = $1
+      WHERE f1.friend_id IS NULL AND f2.user_id IS NULL AND u.id != $1;
+    `,
+      [userId]
+    );
+    return rows;
+  } catch (error) {
+    console.log(error);
+  }
+};
 
 // get all friends of a user
 const fetchFriendsById = async (userId) => {
   try {
     const { rows } = await client.query(
       `
-      SELECT users.username FROM users
+      SELECT users.username, users.id FROM users
       INNER JOIN friends ON users.id = friends.friend_id
       WHERE friends.user_id = $1
       UNION
-      SELECT users.username FROM users
+      SELECT users.username, users.id FROM users
       INNER JOIN friends ON users.id = friends.user_id
       WHERE friends.friend_id = $1
     `,
@@ -82,7 +100,7 @@ WHERE reviews.user_id IN (
 
 const deleteFriend = async (userId, friendId) => {
   try {
-    const SQL = `DELETE FROM friends WHERE user_id = $1 OR friend_id = $1`;
+    const SQL = `DELETE FROM friends WHERE (user_id = $1 and friend_id = $2) OR (user_id = $2 and friend_id = $1)`;
     const { rows } = await client.query(SQL, [userId, friendId]);
     return rows;
   } catch (error) {
@@ -95,4 +113,5 @@ module.exports = {
   fetchFriendsById,
   deleteFriend,
   fetchReviewsByFriends,
+  fetchAvailableFriends,
 };
